feat(client/user-fridge): update imports to use model  for User and Fridge; enhance main page with user information display and user list for fridge information display
