+++ client/src/app/models/fridge.model.ts
@@ -1,0 +1,8 @@
+
+
+export interface Fridge {
+  id: string;
+  name: string;
+  description: string;
+  users: { id: string; name: string }[];
+}
