+++ client/src/app/pages/login-page/subscribe-page/subscribe-page.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Router } from '@angular/router';
 import { AuthService } from '../../../services/auth.service';
-import { User } from '../../../services/auth.service';
+import { User } from '../../../models/user.model';
 import { loadStripe } from '@stripe/stripe-js';
 
 @Component({
